requirements.txt
Flask==2.3.3
google-generativeai==0.8.5
gunicorn==20.1.0